Move datetime mocking to conftest.py
